+++ src/renderer/components/states/resetPassword/ResetPasswordContent.tsx
@@ -1,286 +1,11 @@
-/*
- * eIsland - A sleek, Apple Dynamic Island inspired floating widget for Windows, built with Electron.
- * https://github.com/JNTMTMTM/eIsland
- *
- * Copyright (C) 2026 JNTMTMTM
- * Copyright (C) 2026 pyisland.com
- *
- * Original author: JNTMTMTM[](https://github.com/JNTMTMTM)
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- */
-
-/**
- * @file ResetPasswordContent.tsx
- * @description 独立重置密码状态界面
- * @author 鸡哥
- */
-
-import { useEffect, useState } from 'react';
 import type { ReactElement } from 'react';
-import { useTranslation } from 'react-i18next';
-import useIslandStore from '../../../store/slices';
-import { sendUserEmailCode } from '../../../api/user/userAccountApi';
-import { runSliderCaptcha } from '../../../utils/sliderCaptcha';
+import { useResetPassword } from './hooks/useResetPassword';
+import { ResetPasswordForm } from './components/ResetPasswordForm';
 import '../../../styles/settings/settings.css';
 import '../../../styles/auth/auth.css';
 
-type FeedbackType = 'success' | 'error' | 'info';
-
-interface Feedback {
-  type: FeedbackType;
-  text: string;
-}
-
-const EMAIL_PATTERN = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-
 /** 独立重置密码状态内容 */
 export function ResetPasswordContent(): ReactElement {
-  const { t } = useTranslation();
-  const { setLogin, setRegister, returnFromAuth } = useIslandStore();
-  const [email, setEmail] = useState('');
-  const [emailCode, setEmailCode] = useState('');
-  const [newPassword, setNewPassword] = useState('');
-  const [confirmPassword, setConfirmPassword] = useState('');
-  const [newPasswordVisible, setNewPasswordVisible] = useState(false);
-  const [confirmPasswordVisible, setConfirmPasswordVisible] = useState(false);
-  const [sendingCode, setSendingCode] = useState(false);
-  const [sendCooldownSeconds, setSendCooldownSeconds] = useState(0);
-  const [submitting, setSubmitting] = useState(false);
-  const [feedback, setFeedback] = useState<Feedback | null>(null);
-
-  useEffect(() => {
-    if (sendCooldownSeconds <= 0) return;
-    const timer = window.setTimeout(() => {
-      setSendCooldownSeconds((v) => (v > 0 ? v - 1 : 0));
-    }, 1000);
-    return () => {
-      window.clearTimeout(timer);
-    };
-  }, [sendCooldownSeconds]);
-
-  const handleSendCode = async (): Promise<void> => {
-    if (sendingCode || sendCooldownSeconds > 0) return;
-    const cleanEmail = email.trim().toLowerCase();
-    if (!EMAIL_PATTERN.test(cleanEmail)) {
-      setFeedback({ type: 'error', text: t('settings.user.feedback.emailInvalid', { defaultValue: '请输入有效邮箱地址' }) });
-      return;
-    }
-    setSendingCode(true);
-    let captchaTicket = '';
-    let captchaRandstr = '';
-    let captchaSign = '';
-    try {
-      const captcha = await runSliderCaptcha(cleanEmail);
-      if (!captcha) {
-        setSendingCode(false);
-        setFeedback({ type: 'error', text: t('settings.user.feedback.captchaCancelled', { defaultValue: '请完成滑块验证后再发送验证码' }) });
-        return;
-      }
-      captchaTicket = captcha.ticket;
-      captchaRandstr = captcha.randstr;
-      captchaSign = captcha.sign;
-    } catch (err) {
-      setSendingCode(false);
-      const msg = err instanceof Error ? err.message : t('settings.user.feedback.emailCodeSendFailed', { defaultValue: '验证码发送失败' });
-      setFeedback({ type: 'error', text: msg });
-      return;
-    }
-    const result = await sendUserEmailCode(cleanEmail, 'RESET_PASSWORD', { ticket: captchaTicket, randstr: captchaRandstr, sign: captchaSign });
-    setSendingCode(false);
-    if (!result.ok) {
-      setFeedback({ type: 'error', text: result.message || t('settings.user.feedback.emailCodeSendFailed', { defaultValue: '验证码发送失败' }) });
-      return;
-    }
-    const cooldown = Math.max(0, Number(result.data?.retryAfterSeconds || 60));
-    if (cooldown > 0) {
-      setSendCooldownSeconds(cooldown);
-    }
-    setFeedback({ type: 'success', text: t('settings.user.feedback.emailCodeSent', { defaultValue: '验证码已发送，请查收邮箱' }) });
-  };
-
-  const handleSubmit = (): void => {
-    if (submitting) return;
-    const cleanEmail = email.trim().toLowerCase();
-    if (!EMAIL_PATTERN.test(cleanEmail)) {
-      setFeedback({ type: 'error', text: t('settings.user.feedback.emailInvalid', { defaultValue: '请输入有效邮箱地址' }) });
-      return;
-    }
-    if (!emailCode.trim()) {
-      setFeedback({ type: 'error', text: t('settings.user.feedback.emailCodeRequired', { defaultValue: '请输入邮箱验证码' }) });
-      return;
-    }
-    if (!newPassword) {
-      setFeedback({ type: 'error', text: t('settings.user.feedback.passwordRequired', { defaultValue: '请输入密码' }) });
-      return;
-    }
-    if (newPassword.length < 8) {
-      setFeedback({ type: 'error', text: t('settings.user.feedback.passwordTooShort', { defaultValue: '密码至少 8 位且包含字母与数字' }) });
-      return;
-    }
-    if (!confirmPassword) {
-      setFeedback({ type: 'error', text: t('settings.user.feedback.confirmPasswordRequired', { defaultValue: '请再次输入密码' }) });
-      return;
-    }
-    if (newPassword !== confirmPassword) {
-      setFeedback({ type: 'error', text: t('settings.user.feedback.passwordNotMatch', { defaultValue: '两次输入的密码不一致' }) });
-      return;
-    }
-    setSubmitting(true);
-    setFeedback({
-      type: 'info',
-      text: t('settings.user.feedback.resetPasswordGuidance', {
-        defaultValue: '验证码校验流程已就绪，请前往用户中心完成密码修改。',
-      }),
-    });
-    setSubmitting(false);
-  };
-
-  const renderFeedback = (): ReactElement | null => {
-    if (!feedback) return null;
-    return <div className={`settings-user-feedback settings-user-feedback--${feedback.type}`}>{feedback.text}</div>;
-  };
-
-  return (
-    <div className="auth-state-content" onClick={(e) => e.stopPropagation()}>
-      <div className="auth-panel">
-        <div className="auth-panel-title">{t('settings.user.auth.resetPassword', { defaultValue: '重置密码' })}</div>
-        <div className="auth-panel-subtitle">
-          {t('settings.user.auth.resetPasswordHint', { defaultValue: '通过邮箱验证码校验身份后，可继续修改登录密码' })}
-        </div>
-
-        <label className="settings-field">
-          <span className="settings-field-label">{t('settings.user.fields.email', { defaultValue: '邮箱' })}</span>
-          <input
-            className="settings-field-input"
-            type="email"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            placeholder="user@example.com"
-          />
-        </label>
-
-        <label className="settings-field">
-          <span className="settings-field-label">{t('settings.user.fields.emailCode', { defaultValue: '邮箱验证码' })}</span>
-          <div className="auth-password-input-wrap">
-            <input
-              className="settings-field-input"
-              value={emailCode}
-              onChange={(e) => setEmailCode(e.target.value)}
-              placeholder={t('settings.user.fields.emailCodePlaceholder', { defaultValue: '请输入邮箱验证码' })}
-            />
-            <button
-              type="button"
-              className="auth-password-toggle"
-              onClick={() => void handleSendCode()}
-              disabled={sendingCode || sendCooldownSeconds > 0}
-            >
-              {sendingCode
-                ? t('settings.user.feedback.emailCodeSending', { defaultValue: '发送中…' })
-                : sendCooldownSeconds > 0
-                  ? t('settings.user.actions.sendCodeCooldown', { defaultValue: '{{seconds}}s后重试', seconds: sendCooldownSeconds })
-                  : t('settings.user.actions.sendCode', { defaultValue: '发送验证码' })}
-            </button>
-          </div>
-        </label>
-
-        <label className="settings-field">
-          <span className="settings-field-label">{t('settings.user.fields.newPassword', { defaultValue: '新密码' })}</span>
-          <div className="auth-password-input-wrap">
-            <input
-              className="settings-field-input"
-              type={newPasswordVisible ? 'text' : 'password'}
-              value={newPassword}
-              onChange={(e) => setNewPassword(e.target.value)}
-              placeholder={t('settings.user.fields.newPasswordPlaceholder', { defaultValue: '留空则不修改，至少 8 位含字母数字' })}
-            />
-            <button
-              type="button"
-              className="auth-password-toggle"
-              onClick={() => setNewPasswordVisible((v) => !v)}
-              aria-label={newPasswordVisible
-                ? t('settings.user.actions.hidePassword', { defaultValue: '隐藏密码' })
-                : t('settings.user.actions.showPassword', { defaultValue: '显示密码' })}
-            >
-              {newPasswordVisible
-                ? t('settings.user.actions.hide', { defaultValue: '隐藏' })
-                : t('settings.user.actions.show', { defaultValue: '显示' })}
-            </button>
-          </div>
-        </label>
-
-        <label className="settings-field">
-          <span className="settings-field-label">{t('settings.user.fields.confirmPassword', { defaultValue: '确认密码' })}</span>
-          <div className="auth-password-input-wrap">
-            <input
-              className="settings-field-input"
-              type={confirmPasswordVisible ? 'text' : 'password'}
-              value={confirmPassword}
-              onChange={(e) => setConfirmPassword(e.target.value)}
-              placeholder={t('settings.user.fields.confirmPasswordPlaceholder', { defaultValue: '请再次输入密码' })}
-            />
-            <button
-              type="button"
-              className="auth-password-toggle"
-              onClick={() => setConfirmPasswordVisible((v) => !v)}
-              aria-label={confirmPasswordVisible
-                ? t('settings.user.actions.hidePassword', { defaultValue: '隐藏密码' })
-                : t('settings.user.actions.showPassword', { defaultValue: '显示密码' })}
-            >
-              {confirmPasswordVisible
-                ? t('settings.user.actions.hide', { defaultValue: '隐藏' })
-                : t('settings.user.actions.show', { defaultValue: '显示' })}
-            </button>
-          </div>
-        </label>
-
-        {renderFeedback()}
-
-        <div className="auth-panel-actions">
-          <button
-            type="button"
-            className="settings-user-primary-btn"
-            onClick={handleSubmit}
-            disabled={submitting}
-          >
-            {submitting
-              ? t('settings.user.feedback.submitting', { defaultValue: '处理中…' })
-              : t('settings.user.auth.resetPasswordBtn', { defaultValue: '重置密码' })}
-          </button>
-          <button
-            type="button"
-            className="settings-user-secondary-btn"
-            onClick={() => setLogin()}
-            disabled={submitting}
-          >
-            {t('settings.user.auth.login', { defaultValue: '登录' })}
-          </button>
-          <button
-            type="button"
-            className="settings-user-secondary-btn"
-            onClick={() => setRegister()}
-            disabled={submitting}
-          >
-            {t('settings.user.auth.register', { defaultValue: '注册' })}
-          </button>
-          <button
-            type="button"
-            className="settings-user-secondary-btn"
-            onClick={returnFromAuth}
-            disabled={submitting}
-          >
-            {t('settings.user.actions.backToCenter', { defaultValue: '返回用户中心' })}
-          </button>
-        </div>
-      </div>
-    </div>
-  );
+  const resetPassword = useResetPassword();
+  return <ResetPasswordForm {...resetPassword} />;
 }
