+++ src/renderer/components/states/resetPassword/config/resetPasswordConfig.ts
@@ -1,0 +1,8 @@
+export type FeedbackType = 'success' | 'error' | 'info';
+
+export interface Feedback {
+  type: FeedbackType;
+  text: string;
+}
+
+export const EMAIL_PATTERN = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
