+++ src/renderer/components/states/resetPassword/index.ts
@@ -1,0 +1,1 @@
+export { ResetPasswordContent } from './ResetPasswordContent';
